More stripe utility moves; more unit tests
